HSEARCH-1614 Adding check for unkown field in MoreLikeThisBuilder.
Also splitting up DSLTest, extracting a dedicated MoreLikeThisTest
